color capture -> ask user -> gray video

+++ sources/ms.sh
@@ -3,17 +3,15 @@
 
 #Functions
 display_usage() {
-	echo -e "Movie digitalizer scipt\n"
-	echo -e "Usage:\n$0 x [-s|n] [-m|c] [-p]\n"
-        echo -e "   x : number of film in format %02d\n"
-        echo -e "  -m : Monochrome images\n"
-        echo -e "  -c : Color images\n"
-        echo -e "  -s : Super film\n"
-        echo -e "  -n : Normal film\n"
+	echo "Movie digitalizer scipt"
+	echo "Usage:\n$0 x [-s|n]"
+        echo "   x : number of film in format %02d"
+        echo "  -s : Super film"
+        echo "  -n : Normal film"
 }
 
 line() {
-        echo -e "------------------------------------\n"
+        echo "------------------------------------"
 }
 # Read my NAS account data (NAS_USER, NAS_PASS)
 . ~/.my-accounts
@@ -32,52 +30,83 @@ fi
 
 # jump to temp folder where images will be stored
 line
-echo -e "----- Jump to ~/temp folder\n"
+echo "----- Jump to ~/temp folder"
 cd ~/temp
 if [ $? -ne 0 ] # cd command faults
 then
-  echo -e "ERROR! Cannot jump to ~/temp folder."
+  echo "ERROR! Cannot jump to ~/temp folder."
   exit
 fi
 if [ "$(ls -A ~/temp)" ] # If temp folder is not empty
 then
-  cd ~/
-  sudo rm -r temp # Delete all images in folder
-  mkdir temp
-  cd ~/temp
+  read -p "There are already images available. Do you want to overwrite them? (y/n)?" choice
+  case "$choice" in
+    y|Y ) NEW_IMAGES=1;;
+    n|N ) ;;
+    * ) echo "invalid";;
+  esac
+  if [ $NEW_IMAGES ]; then
+    cd ~/
+    sudo rm -r temp # Delete all images in folder
+    mkdir temp
+    cd ~/temp
+  fi
+else # If temp folder is empty
+  NEW_IMAGES=1
 fi
 
 # Shoot slides of movie. The sript exits at the end of movie automaticly
-line
-echo -e "----- Shooting slides ($2 $3)\n"
-sudo python ~/movscan/sources/py/movscan.py -p $2 $3
-if [ $? -ne 0 ]; then
-  echo -e "ERROR! Cannot take images."
-  exit
+if [ $NEW_IMAGES ]; then
+  line
+  echo "----- Shooting slides ($2 $3)"
+  sudo python ~/movscan/sources/py/movscan.py -p $2
+  if [ $? -ne 0 ]; then
+    echo "ERROR! Cannot take images."
+    exit
+  fi
 fi
+
+# Ask user whether the film was color or grayscale
+read -p "Was the film color or grayscale? (c/g)?" choice
+case "$choice" in
+  c|C ) echo "color"; FILM_COLOR=1;;
+  g|G ) echo "grayscale"; FILM_GRAYSCALE=1;;
+  * ) echo "invalid answer"; exit;;
+esac
+
 
 # Create video from images
 line
-echo -e "----- Creating 15fps video by concatenating images\n"
+echo "----- Creating 15fps video by concatenating images"
 if [ -f ~/$FILENAME ] # if there is video file with same name
 then
   rm ~/$FILENAME # delete it
 fi
-avconv -g 0 -r 15 -f image2 -i image%05d.jpg -qscale 1 -b 10M -preset slower ~/$FILENAME
+
+if [ $FILM_COLOR ];
+then
+  avconv -g 0 -r 15 -f image2 -i image%05d.jpg -qscale 1 -b 15M -preset slower ~/$FILENAME
+fi
+
+if [ $FILM_GRAYSCALE ];
+then
+  avconv -g 0 -r 15 -f image2 -i image%05d.jpg -flags gray -qscale 1 -b 15M -preset slower ~/$FILENAME
+fi
+
 if [ $? -ne 0 ]; then
-  echo -e "ERROR! Cannot make video."
+  echo "ERROR! Cannot make video."
   exit
 fi
 
 # delete images
 line
-echo -e "----- Delete images\n"
+echo "----- Delete images"
 cd ~/
 sudo rm -r temp # Delete all images in folder
 mkdir temp
 cd ~/temp
 if [ $? -ne 0 ]; then
-  echo -e "  WARNING! Cannot delete images."
+  echo "  WARNING! Cannot delete images."
 fi
 
 # Check internet connection (Thanks to Jesse: http://stackoverflow.com/users/2083761/jesse)
@@ -88,12 +117,12 @@ done
 
 # copy video to NAS if possible
 line
-echo -e "----- Copy video to NAS\n"
+echo "----- Copy video to NAS"
 if ! [ $OnLine ]; then echo "  There is not LAN connection. NAS save is skipped."; fi
 if [ $OnLine ]; then
   if ! mountpoint -q /home/pi/nas
   then
-    echo -e "  - Mount NAS\n"
+    echo "  - Mount NAS"
     sudo mount -t cifs $NASDRIVE /home/pi/nas -o username=$NAS_USER,password=$NAS_PASS
     # return value must not be tested here, because must continue even if mount has failed
   fi
@@ -101,7 +130,7 @@ if [ $OnLine ]; then
   if mountpoint -q /home/pi/nas # If mount is now visible
   then
     # copy video to NAS
-    echo -e "  - Copy video\n"
+    echo "  - Copy video"
     if [ ! -d $NASPATH/$FOLDERNAME ] #if folder does not exist on NAS
     then
       mkdir $NASPATH/$FOLDERNAME # Create it
@@ -125,7 +154,7 @@ fi
 
 # upload video to YouTube
 line
-echo -e "----- Upload video to YouTube\n"
+echo "----- Upload video to YouTube"
 
 if ! [ $OnLine ]; then echo "  There is not Internet connection. YouTube upload is skipped."; fi
 if [ $OnLine ]; then
@@ -133,7 +162,7 @@ if [ $OnLine ]; then
   then
     rm ~/youtube-link # Delete it
     if [ $? -ne 0 ]; then
-      echo -e "  WARNING! Cannot delete link."
+      echo "  WARNING! Cannot delete link."
     fi
   fi
   if [ ! -f ~/youtube-link ] # if there is no link
@@ -142,10 +171,10 @@ if [ $OnLine ]; then
     if [ -f $NASPATH/$FOLDERNAME/$FOLDERNAME.txt ] # if description is alredy available
     then
       DESCRIPTION=$(< $NASPATH/$FOLDERNAME/$FOLDERNAME.txt)
-      echo -e "  Description available."
+      echo "  Description available."
     else
       DESCRIPTION=$FOLDERNAME
-      echo -e "  Description missing."
+      echo "  Description missing."
     fi
     youtube-upload --email=$GMAIL --password=$GPASS --unlisted --title="$FOLDERNAME" --description="$DESCRIPTION" --category=People --keywords="8mm, film, movie, cine-projector, raspberry pi, raspicam, scan, digitalize" ~/$FILENAME >~/youtube-link
     if [ $? -eq 0 ]; then
@@ -155,26 +184,26 @@ if [ $OnLine ]; then
 fi
 
 line
-echo -e "----- Delete video\n"
+echo "----- Delete video"
 # delete video if it is stored on NAS
 if [ $NASSAVE ]; then
   sudo rm ~/$FILENAME
 fi
 if ! [ $NASSAVE ]; then
-  echo -e "  ~/$FILENAME is not deleted, because NAS save was not successfull.\n"
+  echo "  ~/$FILENAME is not deleted, because NAS save was not successfull."
 fi
 
 # Send report mail if it was succesfully uploaded to YouTube
 line
-echo -e "----- Send report e-mail ($GTO)\n"
+echo "----- Send report e-mail ($GTO)"
 if [ $YOUTUBE ]; then
   python /home/pi/movscan/sources/py/sendmail.py $GPASS "$GTO" $FOLDERNAME $(<~/youtube-link) "$DESCRIPTION"
 fi
 if ! [ $YOUTUBE ]; then
-  echo -e "  Email has not sent, because YouTube upload was not successfull.\n"
+  echo "  Email has not sent, because YouTube upload was not successfull."
 fi
 
 if [ $? -eq 0 ]; then
-  echo -e "Hurray, Finished!\n"
+  echo "Hurray, Finished!"
 fi
